+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- syntax check
+- code completion
+- navigate to definition
+
 ## 0.4.17 2019-02-09
 
 ### Fixed